+++ devel/CHANGELOG.md
@@ -10,6 +10,13 @@
 - [trvl]: trivial modifications;
 - [warn]: fixing compiler warning messages;
 
+## 2020-05-31:
+
+- CIO:
+  - introducing `Terminal` type instead of `Console` target,
+- CSDBNS: [adpt],
+- SCLM: [adpt],
+
 ## 2020-05-26:
 
 - MTX: